Separate envvar util header
